fix(quality): reduce Program.cs complexity, re-promote CA1502 to error
Extract the branchy DI-registration blocks (configuration-driven auth, demo
notifications, MassTransit messaging) from Program.cs into ProgramRegistration
extension methods, dropping the top-level <Main>$ cyclomatic complexity below
the CA1502 threshold of 10.

Re-promote CA1502 to an error: remove it from WarningsNotAsErrors and set
.editorconfig severity = error; verify-gates now also asserts the CA1502 fixture
fails under FlowHub's resolved config.

Behaviour preserved — registration order unchanged; integration tests
(WebApplicationFactory<Program>, 21) and core unit tests green.

Closes #97
