+++ source/FlowHub.Web/Program.cs
@@ -3,14 +3,9 @@ using FlowHub.Api;
 using FlowHub.Api.Endpoints;
 using FlowHub.Persistence;
 using FlowHub.Skills;
-using FlowHub.Web.Auth;
+using FlowHub.Web;
 using FlowHub.Web.Components;
-using FlowHub.Web.Notifications;
-using FlowHub.Web.Pipeline;
 using FlowHub.Web.Testing;
-using MassTransit;
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
 using MudBlazor.Services;
 using OpenTelemetry.Metrics;
 using Scalar.AspNetCore;
@@ -25,56 +20,8 @@ builder.Services
 // MudBlazor — only component library per CLAUDE.md.
 builder.Services.AddMudServices();
 
-// Auth mode is driven by configuration, not environment name (12-Factor III).
-// Set Auth__OIDC__Authority + Auth__OIDC__ClientId + Auth__OIDC__ClientSecret for real OIDC.
-// Omit all Auth__OIDC__* vars to activate DemoAuthHandler (any environment).
-if (builder.Configuration["Auth:OIDC:Authority"] is { Length: > 0 } oidcAuthority)
-{
-    var clientId = builder.Configuration["Auth:OIDC:ClientId"]
-        ?? throw new InvalidOperationException("Auth:OIDC:ClientId is required when Auth:OIDC:Authority is set.");
-    var clientSecret = builder.Configuration["Auth:OIDC:ClientSecret"]
-        ?? throw new InvalidOperationException("Auth:OIDC:ClientSecret is required when Auth:OIDC:Authority is set.");
-
-    // Policy scheme dispatches per request: bearer token → JwtBearer (API clients get 401),
-    // anything else → cookie+OIDC (browser flow gets 302 redirect).
-    const string SmartScheme = "smart";
-    builder.Services
-        .AddAuthentication(options =>
-        {
-            options.DefaultScheme = SmartScheme;
-            options.DefaultChallengeScheme = SmartScheme;
-        })
-        .AddPolicyScheme(SmartScheme, SmartScheme, options =>
-        {
-            options.ForwardDefaultSelector = ctx =>
-                ctx.Request.Headers.Authorization.ToString()
-                    .StartsWith("Bearer ", StringComparison.OrdinalIgnoreCase)
-                    ? JwtBearerDefaults.AuthenticationScheme
-                    : Microsoft.AspNetCore.Authentication.Cookies.CookieAuthenticationDefaults.AuthenticationScheme;
-        })
-        .AddCookie()
-        .AddJwtBearer(options =>
-        {
-            options.Authority = oidcAuthority;
-            options.Audience = clientId;
-            options.RequireHttpsMetadata = !builder.Environment.IsDevelopment();
-        })
-        .AddOpenIdConnect(options =>
-        {
-            options.Authority = oidcAuthority;
-            options.ClientId = clientId;
-            options.ClientSecret = clientSecret;
-            options.ResponseType = "code";
-            options.SaveTokens = true;
-        });
-}
-else
-{
-    builder.Services
-        .AddAuthentication(DemoAuthHandler.SchemeName)
-        .AddScheme<AuthenticationSchemeOptions, DemoAuthHandler>(DemoAuthHandler.SchemeName, _ => { });
-}
-
+// Configuration-driven auth (real OIDC vs demo handler) — see ProgramRegistration.
+builder.AddFlowHubAuthentication();
 builder.Services.AddAuthorization();
 builder.Services.AddCascadingAuthenticationState();
 
@@ -119,62 +66,11 @@ builder.Services.AddFlowHubSkills(builder.Configuration);
 builder.Services.Configure<FlowHub.Web.Demo.DemoTraceOptions>(
     builder.Configuration.GetSection(FlowHub.Web.Demo.DemoTraceOptions.SectionName));
 
-// Demo-only operator notifications → ntfy.sh. Dormant unless Demo:Notify:Ntfy:BaseUrl + Topic
-// are set, so the normal app / agent-dev trial never call out. ntfy → Telegram is bridged at
-// the ntfy layer, keeping FlowHub transport-agnostic.
-var demoNotify = builder.Configuration.GetSection(DemoNotifyOptions.SectionName).Get<DemoNotifyOptions>()
-    ?? new DemoNotifyOptions();
-var demoNotifyEnabled = demoNotify.IsConfigured;
-if (demoNotifyEnabled)
-{
-    builder.Services.Configure<DemoNotifyOptions>(builder.Configuration.GetSection(DemoNotifyOptions.SectionName));
-    builder.Services.AddHttpClient<ICaptureNotifier, NtfyCaptureNotifier>(client =>
-    {
-        client.BaseAddress = new Uri(demoNotify.BaseUrl!.TrimEnd('/') + "/");
-        client.Timeout = TimeSpan.FromSeconds(10);
-    });
-}
+// Demo-only operator notifications → ntfy.sh (dormant unless configured) — see ProgramRegistration.
+var demoNotifyEnabled = builder.AddFlowHubDemoNotifications();
 
-// Block 3 Slice B — MassTransit pipeline.
-builder.Services.AddMassTransit(x =>
-{
-    x.SetKebabCaseEndpointNameFormatter();
-
-    x.AddConsumer<CaptureEnrichmentConsumer>(c =>
-        c.UseMessageRetry(r => r.Intervals(100, 500)));
-
-    x.AddConsumer<CaptureEmbeddingConsumer>(c =>
-        c.UseMessageRetry(r => r.Intervals(500, 2000, 5000)));
-
-    x.AddConsumer<SkillRoutingConsumer>(c =>
-        c.UseMessageRetry(r => r.Intervals(500, 2000, 5000)));
-
-    // No retry policy — fault observer is best-effort per spec D5
-    // (recursive retry on Fault<T> would loop forever).
-    x.AddConsumer<LifecycleFaultObserver>();
-
-    // Demo-only: announce new captures to ntfy.sh (registered only when configured).
-    if (demoNotifyEnabled)
-        x.AddConsumer<CaptureNotificationConsumer>(c =>
-            c.UseMessageRetry(r => r.Intervals(500, 2000)));
-
-    if (string.Equals(builder.Configuration["Bus:Transport"], "RabbitMq", StringComparison.OrdinalIgnoreCase))
-    {
-        x.UsingRabbitMq((ctx, cfg) =>
-        {
-            cfg.Host(builder.Configuration["Bus:RabbitMq:Host"], "/", h =>
-            {
-                h.Username(builder.Configuration["Bus:RabbitMq:Username"] ?? "guest");
-                h.Password(builder.Configuration["Bus:RabbitMq:Password"] ?? "guest");
-            });
-            cfg.ConfigureEndpoints(ctx);
-        });
-    }
-    else
-    {
-        x.UsingInMemory((ctx, cfg) => cfg.ConfigureEndpoints(ctx));
-    }
-});
+// Block 3 Slice B — MassTransit pipeline (consumers, retries, transport) — see ProgramRegistration.
+builder.AddFlowHubMessaging(demoNotifyEnabled);
 
 // Block 3 Slice A — REST API surface for non-UI consumers.
 builder.Services.AddFlowHubApi();
