+++ tools/verify-gates.sh
@@ -9,11 +9,11 @@
 # *inside* FlowHub's tree (so they inherit FlowHub's Directory.Build.props /
 # .editorconfig) and asserts each ENFORCED gate fires.
 #
-# Phase-1 note (issue #94): CA1502/CA1506 are demoted to warnings here (debt,
-# #95/#97), so this checks the gates FlowHub currently ENFORCES: IDE0005 (build
-# error) and the method-size check. Re-add CA1502/CA1506 below when Phase-2
-# re-promotes them. (The method-size action step is deferred on Linux —
-# agent-pipeline#91 — but the check script itself is Linux-safe and verified here.)
+# Enforced gates checked here: CA1502 (re-promoted in #97), IDE0005, and the
+# method-size check. CA1506 is still demoted to a warning (debt #95) so it is not
+# asserted as a build-failure yet — add it below when #95 re-promotes it. (The
+# method-size action step is deferred on Linux — agent-pipeline#91 — but the
+# check script itself is Linux-safe and verified here.)
 set -euo pipefail
 IFS=$'\n\t'
 
@@ -42,6 +42,17 @@ if dotnet build "${SCRATCH}/ide/UnusedUsing.csproj" -c Release --nologo -v q >/d
   fail=1
 else
   printf 'OK: IDE0005 fired\n'
+fi
+
+# --- CA1502: re-promoted to an error in #97 ----------------------------------
+note "CA1502 fixture must FAIL the build under FlowHub's resolved config"
+fetch "gate-tests/build-failures/Complexity/Complexity.cs"     "ca1502/Complexity.cs"
+fetch "gate-tests/build-failures/Complexity/Complexity.csproj" "ca1502/Complexity.csproj"
+if dotnet build "${SCRATCH}/ca1502/Complexity.csproj" -c Release --nologo -v q >/dev/null 2>&1; then
+  printf '::error::CA1502 did NOT fail the build — the gate is dead in FlowHub\n'
+  fail=1
+else
+  printf 'OK: CA1502 fired\n'
 fi
 
 # --- Method-size check (Linux-safe; independent of the Metrics.exe generator) -
